re-plug the old strategies test. Still a lot of refactoring to do

+++ sieve/ecm/facul_method.cpp
@@ -31,8 +31,8 @@ facul_method::facul_method(parameters const & p, const int verbose)
     /* we don't need to store B1 and B2 in the method object. They all go
      * in the bytecode anyway
      */
-    int const B1 = p.B1;
-    int const B2 = p.B2;
+    unsigned long const B1 = p.B1;
+    unsigned long const B2 = p.B2;
 
     switch(method) {
         case PP1_27_METHOD:
